move query specific code to QueryExecutor
