+++ modules/engine/src/main/scala/com/gsk/kg/engine/Engine.scala
@@ -67,40 +67,8 @@ object Engine {
     eval(query.r)
       .runA(dataframe)
       .map(_.dataframe)
-      .map(qqq(query))
+      .map(QueryExecutor.execute(query))
   }
-
-  private def qqq(
-      query: Query
-  )(df: DataFrame)(implicit sc: SQLContext): DataFrame =
-    query match {
-      case Construct(vars, bgp, r) =>
-        import sc.implicits._
-        val acc = List.empty[(String, String, String)].toDF("s", "p", "o")
-
-        bgp.triples
-          .map({ triple =>
-            import org.apache.spark.sql.functions._
-
-            val cols = (triple.getVariables ++ triple.getPredicates)
-
-            cols
-              .foldLeft(df)({
-                case (df, (sv, pos)) =>
-                  if (df.columns.contains(sv.s)) {
-                    df.withColumnRenamed(sv.s, pos)
-                  } else {
-                    df.withColumn(pos, lit(sv.s))
-                  }
-              })
-              .select("s", "p", "o")
-          })
-          .foldLeft(acc) { (acc, other) =>
-            acc.union(other)
-          }
-
-      case _ => df
-    }
 
   private def evaluateBGPF(
       triples: Seq[Expr.Triple]
